Change the Sink[Json] to be a Flow[Json ~> SierraRecord]
